feat(modif):modifcation in controller and service
