+++ src/main/resources/application.yml
@@ -13,3 +13,8 @@ springdoc:
     path: /docs
 
 
+user:
+  service:
+    host: ${USER_SERVICE_HOST:my-spring-boot-app}
+    port: ${USER_SERVICE_PORT:9090}
+    path: ${USER_SERVICE_PATH:/user}
